Changed int's to short int's for the multiindex mapping to save space. Worked on a bit on Fourier expansion. Now need to work on the error bound.

+++ fastlib2/mlpack/series_expansion/inverse_pow_dist_kernel_aux.h
@@ -11,8 +11,8 @@ class InversePowDistGradientKernelAux {
  private:
 
   void SubFrom_(index_t dimension, int decrement,
-		const ArrayList<int> &subtract_from, 
-		ArrayList<int> &result) const {
+		const ArrayList<short int> &subtract_from, 
+		ArrayList<short int> &result) const {
     
     for(index_t d = 0; d < subtract_from.size(); d++) {
       if(d == dimension) {
@@ -70,7 +70,7 @@ class InversePowDistGradientKernelAux {
 
     // Temporary variable to look for arithmetic operations on
     // multiindex.
-    ArrayList<int> tmp_multiindex;
+    ArrayList<short int> tmp_multiindex;
     tmp_multiindex.Init(sea_.get_dimension());
 
     for(index_t i = 0; i < derivative_map->n_rows(); i++) {
@@ -79,7 +79,7 @@ class InversePowDistGradientKernelAux {
       double contribution = 0;
 
       // Retrieve the multiindex mapping.
-      const ArrayList<int> &multiindex = sea_.get_multiindex(i);
+      const ArrayList<short int> &multiindex = sea_.get_multiindex(i);
       
       // $D_{x}^{0} \phi_{\nu, d}(x)$ should be computed normally.
       if(i == 0) {
@@ -133,7 +133,7 @@ class InversePowDistGradientKernelAux {
     for(index_t i = 1; i < derivative_map->n_rows(); i++) {
 
       // Retrieve the multiindex mapping.
-      const ArrayList<int> &multiindex = sea_.get_multiindex(i);
+      const ArrayList<short int> &multiindex = sea_.get_multiindex(i);
 
       // The sum of the indices.
       index_t sum_of_indices = 0;
@@ -148,7 +148,7 @@ class InversePowDistGradientKernelAux {
   }
   
   double ComputePartialDerivative(const Matrix &derivative_map,
-				  const ArrayList<int> &mapping) const {
+				  const ArrayList<short int> &mapping) const {
     
     return derivative_map.get(sea_.ComputeMultiindexPosition(mapping), 0);
   }
@@ -162,8 +162,8 @@ class InversePowDistKernelAux {
 
  private:
   void SubFrom_(index_t dimension, int decrement,
-		const ArrayList<int> &subtract_from, 
-		ArrayList<int> &result) const {
+		const ArrayList<short int> &subtract_from, 
+		ArrayList<short int> &result) const {
     
     for(index_t d = 0; d < subtract_from.size(); d++) {
       if(d == dimension) {
@@ -215,7 +215,7 @@ class InversePowDistKernelAux {
 
     // Temporary variable to look for arithmetic operations on
     // multiindex.
-    ArrayList<int> tmp_multiindex;
+    ArrayList<short int> tmp_multiindex;
     tmp_multiindex.Init(sea_.get_dimension());
 
     // Get the inverse multiindex factorial factors.
@@ -228,7 +228,7 @@ class InversePowDistKernelAux {
       double contribution = 0;
 
       // Retrieve the multiindex mapping.
-      const ArrayList<int> &multiindex = sea_.get_multiindex(i);
+      const ArrayList<short int> &multiindex = sea_.get_multiindex(i);
       
       // $D_{x}^{0} \phi_{\nu, d}(x)$ should be computed normally.
       if(i == 0) {
@@ -290,7 +290,7 @@ class InversePowDistKernelAux {
     for(index_t i = 1; i < derivative_map->n_rows(); i++) {
 
       // Retrieve the multiindex mapping.
-      const ArrayList<int> &multiindex = sea_.get_multiindex(i);
+      const ArrayList<short int> &multiindex = sea_.get_multiindex(i);
 
       // The sum of the indices.
       index_t sum_of_indices = 0;
@@ -305,7 +305,7 @@ class InversePowDistKernelAux {
   }
   
   double ComputePartialDerivative(const Matrix &derivative_map,
-				  const ArrayList<int> &mapping) const {
+				  const ArrayList<short int> &mapping) const {
     
     return derivative_map.get(sea_.ComputeMultiindexPosition(mapping), 0);
   }
